cleanup + ensureChar Exists
